+++ StoreRotationConfig/Plugin.cs
@@ -10,10 +10,10 @@ namespace StoreRotationConfig
     ///     Simple mod that adds configurability to the number of items that show up in the store every week.
     /// </summary>
     [BepInPlugin(GUID, PLUGIN_NAME, VERSION)]
-    [BepInDependency("com.sigurd.csync", "4.1.0")]
+    [BepInDependency("com.sigurd.csync", "5.0.0")]
     public class Plugin : BaseUnityPlugin
     {
-        internal const string GUID = "pacoito.StoreRotationConfig", PLUGIN_NAME = "StoreRotationConfig", VERSION = "2.0.1";
+        internal const string GUID = "pacoito.StoreRotationConfig", PLUGIN_NAME = "StoreRotationConfig", VERSION = "2.1.0";
         internal static ManualLogSource StaticLogger { get; private set; }
 
         /// <summary>